LUN-3: Move Interfaces from index.ts
